+++ src/candlewick/core/math_types.h
@@ -170,10 +170,4 @@ namespace math {
   }
 } // namespace math
 
-class Mesh;
-
-/// \brief Intermediary argument type for shadow-casting or opaque objects. For
-/// use in depth or light pre-passes.
-using OpaqueCastable = std::tuple<const Mesh &, Mat4f>;
-
 } // namespace candlewick